requirements.txt
docker
google-genai